Success! Now there is a separate operation type (coalescing) and everything seems to work fine! Also fixed link and expression types, since they were jumping to the string on a loop instead or respective labels.

+++ common_types.h
@@ -237,7 +237,7 @@ struct Span : Array<uint8> {
 		max_index -= shift;
 	}
 
-	inline void move_relative(size_t position, size_t size, size_t shift) {
+	inline void move_relative(size_t position, size_t size, int64 shift) {
 		memcpy(content + position + shift, content + position, size);
 	}
